+++ src/utils/formatErrorMessage.ts
@@ -1,3 +1,5 @@
+import isEmpty from './isEmpty';
+
 /**
  * formatErrorMessage('${name} is a required field', {name: 'email'});
  * output: 'email is a required field'
@@ -5,7 +7,7 @@
 export default function formatErrorMessage<E>(errorMessage?: string | E, params?: any) {
   if (typeof errorMessage === 'string') {
     return errorMessage.replace(/\$\{\s*(\w+)\s*\}/g, (_, key) => {
-      return params?.[key] || `[${key}]`;
+      return isEmpty(params?.[key]) ? `[${key}]` : params?.[key];
     });
   }