+++ test/MixedTypeSpec.js
@@ -1,7 +1,7 @@
 /* eslint-disable @typescript-eslint/no-var-requires */
-require('chai').should();
-
+const chai = require('chai');
 const schema = require('../src');
+chai.should();
 const { StringType, SchemaModel, NumberType, ArrayType, MixedType } = schema;
 
 describe('#MixedType', () => {
@@ -400,5 +400,64 @@ describe('#MixedType', () => {
     checkResult.age.hasError.should.equal(true);
     checkResult.contact.hasError.should.equal(true);
     checkResult.contact.errorMessage.should.equal('error2');
+  });
+
+  it('should error when an async rule is executed by the sync validator', () => {
+    const m = MixedType().addRule(async () => {
+      return true;
+    }, 'An async error');
+    let err;
+    try {
+      m.check({});
+    } catch (e) {
+      err = e;
+    }
+    chai
+      .expect(err?.message)
+      .to.eql('synchronous validator had an async result, you should probably call "checkAsync()"');
+  });
+  it('Should be able to check by `checkAsync` with `addAsyncRule`', done => {
+    const type = MixedType()
+      .addAsyncRule(v => {
+        return new Promise(resolve => {
+          setTimeout(() => {
+            if (typeof v === 'number') {
+              resolve(true);
+            } else {
+              resolve(false);
+            }
+          }, 500);
+        });
+      }, 'error1')
+      .isRequired('error2');
+
+    Promise.all([type.checkAsync(''), type.checkAsync('1'), type.checkAsync(1)]).then(res => {
+      if (res[0].hasError && res[1].hasError && !res[2].hasError) {
+        done();
+      }
+    });
+  });
+  it('Should be able to check by `check` with `addAsyncRule` and skip the async ', done => {
+    let called = false;
+    const type = MixedType()
+      .addRule(v => {
+        return typeof v === 'number';
+      }, 'This is not async')
+      .addAsyncRule(async () => {
+        called = true;
+        return false;
+      }, 'error1')
+      .isRequired('error2');
+    setTimeout(() => {
+      try {
+        chai.expect(called).to.eq(false);
+        chai.expect(type.check('').hasError).to.eq(true);
+        chai.expect(type.check('1').hasError).to.eq(true);
+        chai.expect(type.check(1).hasError).to.eq(false);
+        done();
+      } catch (e) {
+        done(e);
+      }
+    }, 100);
   });
 });